the circle is complete! At last the gate will be opened!

+++ WizardWar/Wizard.h
@@ -33,7 +33,7 @@ typedef enum WizardStatus {
 @property (nonatomic, strong) NSString * wizardType;
 @property (nonatomic, strong) PlayerEffect * effect; // current effect applied
 @property (nonatomic) NSInteger effectStartTick;
-@property (nonatomic) NSInteger altitude; // how high you are. 0 = normal
+@property (nonatomic) float altitude; // how high you are. 0 = normal
 @property (readonly) NSInteger direction;
 @property (nonatomic) NSUInteger colorRGB; // hex rgb value
 @property (nonatomic, strong) UIColor * color;
